+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#569](https://github.com/xmake-io/xmake/pull/569): Add c++ modules build rules
 * Add `xmake project -k xmakefile` generator
+* [620](https://github.com/xmake-io/xmake/issues/620): Add global `~/.xmakerc.lua` for all projects.
 
 ### Change
 
@@ -653,6 +654,7 @@
 
 * [#569](https://github.com/xmake-io/xmake/pull/569): 增加对c++模块的实验性支持
 * 添加`xmake project -k xmakefile`生成器
+* [620](https://github.com/xmake-io/xmake/issues/620): 添加全局`~/.xmakerc.lua`配置文件，对所有本地工程生效.
 
 ### 改进